Renamed admin-api to admin-endpoint. Moved support resources from admin-endpoint to new module - customer-support-endpoint. Moved CORS from customer-endpoint to jersey module. Added env and deployment fields to slack alerts.
