+++ .circleci/prime-canary-values.yaml
@@ -31,6 +31,8 @@ prime:
     FIREBASE_ROOT_PATH: v2
     NEO4J_HOST: neo4j-neo4j.neo4j.svc.cluster.local
     SLACK_CHANNEL: prime-alerts
+    SLACK_ENV: Production
+    SLACK_DEPLOYMENT: Canary
     DATA_TRAFFIC_TOPIC: data-traffic
     PURCHASE_INFO_TOPIC: purchase-info
     ACTIVE_USERS_TOPIC: active-users
