--- admin-api/src/main/kotlin/org/ostelco/prime/admin/api/Model.kt
@@ -1,9 +1,0 @@
-package org.ostelco.prime.admin.api
-
-import org.ostelco.prime.model.Plan
-import org.ostelco.prime.model.Product
-
-data class CreatePlanRequest(
-        val plan: Plan,
-        val stripeProductName: String,
-        val planProduct: Product)
