--- admin-api/src/main/resources/META-INF/services/org.ostelco.prime.module.PrimeModule
@@ -1,1 +1,0 @@
-org.ostelco.prime.admin.api.AdminModule
