+++ admin-endpoint/src/main/kotlin/org/ostelco/prime/admin/Model.kt
@@ -1,0 +1,9 @@
+package org.ostelco.prime.admin
+
+import org.ostelco.prime.model.Plan
+import org.ostelco.prime.model.Product
+
+data class CreatePlanRequest(
+        val plan: Plan,
+        val stripeProductName: String,
+        val planProduct: Product)
