+++ admin-endpoint/src/main/resources/META-INF/services/io.dropwizard.jackson.Discoverable
