+++ admin-endpoint/src/test/resources/sample-offer-legacy.yaml
