+++ admin-endpoint/src/test/resources/sample-offer-only.yaml
