+++ admin-endpoint/src/test/resources/sample-offer-products-segments.yaml
