+++ customer-endpoint/src/main/kotlin/org/ostelco/prime/customer/endpoint/CustomerEndpointModule.kt
@@ -2,7 +2,6 @@ package org.ostelco.prime.customer.endpoint
 
 import com.fasterxml.jackson.annotation.JsonTypeName
 import io.dropwizard.setup.Environment
-import org.eclipse.jetty.servlets.CrossOriginFilter
 import org.ostelco.prime.customer.endpoint.metrics.reportMetricsAtStartUp
 import org.ostelco.prime.customer.endpoint.resources.ApplicationTokenResource
 import org.ostelco.prime.customer.endpoint.resources.BundlesResource
@@ -16,8 +15,6 @@ import org.ostelco.prime.customer.endpoint.resources.ReferralResource
 import org.ostelco.prime.customer.endpoint.resources.RegionsResource
 import org.ostelco.prime.customer.endpoint.store.SubscriberDAOImpl
 import org.ostelco.prime.module.PrimeModule
-import java.util.*
-import javax.servlet.DispatcherType
 
 
 /**
@@ -28,16 +25,6 @@ import javax.servlet.DispatcherType
 class CustomerEndpointModule : PrimeModule {
 
     override fun init(env: Environment) {
-
-        // Allow CORS
-        val corsFilterRegistration = env.servlets().addFilter("CORS", CrossOriginFilter::class.java)
-        // Configure CORS parameters
-        corsFilterRegistration.setInitParameter("allowedOrigins", "*")
-        corsFilterRegistration.setInitParameter("allowedHeaders",
-                "Cache-Control,If-Modified-Since,Pragma,Content-Type,Authorization,X-Requested-With,Content-Length,Accept,Origin")
-        corsFilterRegistration.setInitParameter("allowedMethods", "OPTIONS,GET,PUT,POST,DELETE,HEAD")
-        corsFilterRegistration.addMappingForUrlPatterns(EnumSet.allOf(DispatcherType::class.java), true, "/*")
-
 
         val dao = SubscriberDAOImpl()
         val jerseyEnv = env.jersey()
