+++ customer-support-endpoint/build.gradle
@@ -1,0 +1,13 @@
+plugins {
+  id "org.jetbrains.kotlin.jvm"
+  id "java-library"
+}
+
+dependencies {
+  implementation project(":prime-modules")
+
+  testImplementation project(":jersey")
+  testImplementation "io.dropwizard:dropwizard-testing:$dropwizardVersion"
+}
+
+apply from: '../gradle/jacoco.gradle'
