+++ customer-support-endpoint/src/main/resources/META-INF/services/io.dropwizard.jackson.Discoverable
@@ -1,0 +1,1 @@
+org.ostelco.prime.module.PrimeModule
