+++ customer-support-endpoint/src/main/resources/META-INF/services/org.ostelco.prime.module.PrimeModule
@@ -1,0 +1,1 @@
+org.ostelco.prime.support.SupportModule
