+++ prime/infra/prime-direct-values.yaml
@@ -28,6 +28,8 @@ prime:
     FIREBASE_ROOT_PATH: dev
     NEO4J_HOST: neo4j-neo4j.neo4j.svc.cluster.local
     SLACK_CHANNEL: prime-alerts
+    SLACK_ENV: Development
+    SLACK_DEPLOYMENT: Direct
     DATA_TRAFFIC_TOPIC: data-traffic
     PURCHASE_INFO_TOPIC: purchase-info
     ACTIVE_USERS_TOPIC: active-users
